+++ modules/runners/lambdas/runners/src/scale-runners/scale-down.test.ts
@@ -316,3 +316,218 @@ describe('scaleDown', () => {
     });
   });
 });
+
+describe('scaleDown ghes', () => {
+  beforeEach(() => {
+    process.env.GITHUB_APP_KEY_BASE64 = 'TEST_CERTIFICATE_DATA';
+    process.env.GITHUB_APP_ID = '1337';
+    process.env.GITHUB_APP_CLIENT_ID = 'TEST_CLIENT_ID';
+    process.env.GITHUB_APP_CLIENT_SECRET = 'TEST_CLIENT_SECRET';
+    process.env.RUNNERS_MAXIMUM_COUNT = '3';
+    process.env.ENVIRONMENT = environment;
+    process.env.MINIMUM_RUNNING_TIME_IN_MINUTES = minimumRunningTimeInMinutes.toString();
+    process.env.GHES_URL = 'https://github.enterprise.something';
+    jest.clearAllMocks();
+    mockOctokit.apps.getOrgInstallation.mockImplementation(() => ({
+      data: {
+        id: 'ORG',
+      },
+    }));
+    mockOctokit.apps.getRepoInstallation.mockImplementation(() => ({
+      data: {
+        id: 'REPO',
+      },
+    }));
+
+    mockOctokit.paginate.mockImplementationOnce(() => {
+      return DEFAULT_REGISTERED_RUNNERS;
+    });
+
+    mockOctokit.actions.deleteSelfHostedRunnerFromRepo.mockImplementation((repo) => {
+      if (repo.runner_id === 103) {
+        throw Error();
+      } else {
+        return { status: 204 };
+      }
+    });
+    mockOctokit.actions.deleteSelfHostedRunnerFromOrg.mockImplementation((repo) => {
+      return repo.runner_id === 103 ? { status: 500 } : { status: 204 };
+    });
+
+    const mockTerminateRunners = mocked(terminateRunner);
+    mockTerminateRunners.mockImplementation(async () => {
+      return;
+    });
+  });
+
+  describe('no runners running', () => {
+    beforeAll(() => {
+      const mockListRunners = mocked(listRunners);
+      mockListRunners.mockImplementation(async () => []);
+    });
+
+    it('No runners for repo.', async () => {
+      process.env.ENABLE_ORGANIZATION_RUNNERS = 'false';
+      process.env.SCALE_DOWN_CONFIG = '[]';
+      await scaleDown();
+      expect(listRunners).toBeCalledWith({
+        environment: environment,
+      });
+      expect(terminateRunner).not;
+      expect(mockOctokit.apps.getRepoInstallation).not;
+    });
+
+    it('No runners for org.', async () => {
+      process.env.ENABLE_ORGANIZATION_RUNNERS = 'true';
+      await scaleDown();
+      expect(listRunners).toBeCalledWith({
+        environment: environment,
+      });
+      expect(terminateRunner).not;
+      expect(mockOctokit.apps.getRepoInstallation).not;
+    });
+  });
+
+  describe('on repo level', () => {
+    beforeAll(() => {
+      process.env.ENABLE_ORGANIZATION_RUNNERS = 'false';
+      process.env.SCALE_DOWN_CONFIG = '[]';
+      const mockListRunners = mocked(listRunners);
+      mockListRunners.mockImplementation(async () => {
+        return DEFAULT_RUNNERS;
+      });
+    });
+
+    it('Terminate 3 of 5 runners for repo.', async () => {
+      await scaleDown();
+      expect(listRunners).toBeCalledWith({
+        environment: environment,
+      });
+
+      expect(mockOctokit.apps.getRepoInstallation).toBeCalled();
+      expect(terminateRunner).toBeCalledTimes(3);
+      for (const toTerminate of DEFAULT_RUNNERS_TO_BE_REMOVED) {
+        expect(terminateRunner).toHaveBeenCalledWith(toTerminate);
+      }
+    });
+  });
+
+  describe('on org level', () => {
+    beforeAll(() => {
+      process.env.ENABLE_ORGANIZATION_RUNNERS = 'true';
+      process.env.SCALE_DOWN_CONFIG = '[]';
+      const mockListRunners = mocked(listRunners);
+      mockListRunners.mockImplementation(async () => {
+        return DEFAULT_RUNNERS;
+      });
+    });
+
+    it('Terminate 3 of 5 runners for org.', async () => {
+      await scaleDown();
+      expect(listRunners).toBeCalledWith({
+        environment: environment,
+      });
+
+      expect(mockOctokit.apps.getOrgInstallation).toBeCalled();
+      expect(terminateRunner).toBeCalledTimes(3);
+      for (const toTerminate of DEFAULT_RUNNERS_TO_BE_REMOVED) {
+        expect(terminateRunner).toHaveBeenCalledWith(toTerminate);
+      }
+    });
+  });
+
+  describe('on repo level', () => {
+    beforeAll(() => {
+      process.env.ENABLE_ORGANIZATION_RUNNERS = 'false';
+      process.env.SCALE_DOWN_CONFIG = '[]';
+      const mockListRunners = mocked(listRunners);
+      mockListRunners.mockImplementation(async () => {
+        return DEFAULT_RUNNERS;
+      });
+    });
+
+    it('Terminate 3 of 5 runners for repo.', async () => {
+      await scaleDown();
+      expect(listRunners).toBeCalledWith({
+        environment: environment,
+      });
+
+      expect(mockOctokit.apps.getRepoInstallation).toBeCalled();
+      expect(terminateRunner).toBeCalledTimes(3);
+      for (const toTerminate of DEFAULT_RUNNERS_TO_BE_REMOVED) {
+        expect(terminateRunner).toHaveBeenCalledWith(toTerminate);
+      }
+    });
+  });
+
+  describe('on org level', () => {
+    beforeAll(() => {
+      process.env.ENABLE_ORGANIZATION_RUNNERS = 'true';
+      process.env.SCALE_DOWN_CONFIG = '[]';
+      const mockListRunners = mocked(listRunners);
+      mockListRunners.mockImplementation(async () => {
+        return DEFAULT_RUNNERS;
+      });
+    });
+
+    it('Terminate 3 of 5 runners for org.', async () => {
+      await scaleDown();
+      expect(listRunners).toBeCalledWith({
+        environment: environment,
+      });
+
+      expect(mockOctokit.apps.getOrgInstallation).toBeCalled();
+      expect(terminateRunner).toBeCalledTimes(3);
+      for (const toTerminate of DEFAULT_RUNNERS_TO_BE_REMOVED) {
+        expect(terminateRunner).toHaveBeenCalledWith(toTerminate);
+      }
+    });
+  });
+
+  describe('Have runners idle.', () => {
+    beforeAll(() => {
+      process.env.SCALE_DOWN_CONFIG = JSON.stringify([
+        {
+          idleCount: 2,
+          cron: '* * * * * *',
+          timeZone: 'Europe/Amsterdam',
+        },
+      ]);
+
+      const mockListRunners = mocked(listRunners);
+      mockListRunners.mockImplementation(async () => {
+        return RUNNERS_WITH_AUTO_SCALING_CONFIG;
+      });
+    });
+
+    it('Terminate 1 of runners for org.', async () => {
+      await scaleDown();
+      process.env.ENABLE_ORGANIZATION_RUNNERS = 'true';
+
+      expect(listRunners).toBeCalledWith({
+        environment: environment,
+      });
+
+      expect(mockOctokit.apps.getOrgInstallation).toBeCalled();
+      expect(terminateRunner).toBeCalledTimes(1);
+      for (const toTerminate of RUNNERS_TO_BE_REMOVED_WITH_AUTO_SCALING_CONFIG) {
+        expect(terminateRunner).toHaveBeenCalledWith(toTerminate);
+      }
+    });
+
+    it('Terminate 1 of runners for repo.', async () => {
+      await scaleDown();
+      process.env.ENABLE_ORGANIZATION_RUNNERS = 'false';
+
+      expect(listRunners).toBeCalledWith({
+        environment: environment,
+      });
+
+      expect(mockOctokit.apps.getOrgInstallation).toBeCalled();
+      expect(terminateRunner).toBeCalledTimes(1);
+      for (const toTerminate of RUNNERS_TO_BE_REMOVED_WITH_AUTO_SCALING_CONFIG) {
+        expect(terminateRunner).toHaveBeenCalledWith(toTerminate);
+      }
+    });
+  });
+});